refactor(ale_claw): group history modules into harness/context/ subpackage
Move context, compaction, token_estimation, transcript, replay into a
context/ subpackage. Behavior-preserving: git mv + import-path updates
(internal non-sibling imports -> .., external importers -> .context.X).
928 passed, 3 skipped.
