+++ ale_run/agents/ale_claw/deployer.py
@@ -59,7 +59,7 @@ from .harness import (
     convert_to_responses_api_items,
 )
 from .harness.agent_loop import has_done_signal
-from .harness.context import DEFAULT_CONTEXT_TOKENS, resolve_context_window
+from .harness.context.context import DEFAULT_CONTEXT_TOKENS, resolve_context_window
 from .harness.model_config import resolve_model
 
 logger = logging.getLogger(__name__)
