+++ ale_run/agents/ale_claw/harness/subagent/subagent_session.py
@@ -37,7 +37,7 @@ from typing import Any
 from ..model_config import ResolvedModel, resolve_model
 from agent.tools.base import BaseTool
 
-from ..context import (
+from ..context.context import (
     ContextOverflowCallback,
     compact_messages,
     is_context_overflow_error,
