+++ tests/ale_claw/test_openclaw_compaction.py
@@ -18,7 +18,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 from ale_run.agents.ale_claw.harness.model_config import HelperTransportDefaults, ModelConfig, register_model_config
 
-from ale_run.agents.ale_claw.harness.context import (
+from ale_run.agents.ale_claw.harness.context.context import (
     BASE_CHUNK_RATIO,
     DEFAULT_SUMMARY_FALLBACK,
     IDENTIFIER_PRESERVATION_INSTRUCTIONS,
@@ -543,12 +543,12 @@ class TestSummarizeWithFallback:
 class TestRemovedSymbols:
     def test_merge_summaries_instructions_removed(self):
         """MERGE_SUMMARIES_INSTRUCTIONS should no longer exist in context module."""
-        import ale_run.agents.ale_claw.harness.context as ctx
+        import ale_run.agents.ale_claw.harness.context.context as ctx
         assert not hasattr(ctx, "MERGE_SUMMARIES_INSTRUCTIONS")
 
     def test_merge_summaries_function_removed(self):
         """_merge_summaries should no longer exist in context module."""
-        import ale_run.agents.ale_claw.harness.context as ctx
+        import ale_run.agents.ale_claw.harness.context.context as ctx
         assert not hasattr(ctx, "_merge_summaries")
